+++ source/core/board/NstBoardBandaiLz93d50.cpp
@@ -155,11 +155,11 @@ namespace Nes
 
 					const byte data[5] =
 					{
-						irq.Connected() ? 0x1 : 0x0,
-						irq.unit.latch >> 0 & 0xFF,
-						irq.unit.latch >> 8 & 0xFF,
-						irq.unit.count >> 0 & 0xFF,
-						irq.unit.count >> 8 & 0xFF
+						static_cast<byte>(irq.Connected() ? 0x1 : 0x0),
+						static_cast<byte>(irq.unit.latch >> 0 & 0xFF),
+						static_cast<byte>(irq.unit.latch >> 8 & 0xFF),
+						static_cast<byte>(irq.unit.count >> 0 & 0xFF),
+						static_cast<byte>(irq.unit.count >> 8 & 0xFF)
 					};
 
 					state.Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End();
