+++ source/core/board/NstBoardBmcGamestarA.cpp
@@ -207,9 +207,9 @@ namespace Nes
 				{
 					const byte data[3] =
 					{
-						cartSwitches ? cartSwitches->GetMode() : 0,
-						regs[0],
-						regs[1]
+						static_cast<byte>(cartSwitches ? cartSwitches->GetMode() : 0),
+						static_cast<byte>(regs[0]),
+						static_cast<byte>(regs[1])
 					};
 
 					state.Begin( AsciiId<'B','G','A'>::V ).Begin( AsciiId<'R','E','G'>::V ).Write( data ).End().End();
