+++ source/core/board/NstBoardBtlSmb3.cpp
@@ -94,9 +94,9 @@ namespace Nes
 				{
 					const byte data[3] =
 					{
-						irq.unit.enabled ? 0x1 : 0x0,
-						irq.unit.count & 0xFF,
-						irq.unit.count >> 8
+						static_cast<byte>(irq.unit.enabled ? 0x1 : 0x0),
+						static_cast<byte>(irq.unit.count & 0xFF),
+						static_cast<byte>(irq.unit.count >> 8)
 					};
 
 					state.Begin( AsciiId<'B','S','3'>::V ).Begin( AsciiId<'I','R','Q'>::V ).Write( data ).End().End();
